refactor: components 분리
